// src/types/actions/boardActions.enum.ts
const enum BoardActions {
	GET_BOARDS = 'get_board',
	CREATE_NEW_BOARD = 'create_new_board',
	CREATE_NEW_BOARD_REQUEST = 'create_new_board_request',
	DELETE_BOARD = 'delete_board',
	FETCH_BOARDS_SEARCH = 'search_board',
	FETCH_BOARDS = 'fetch_boards',
	FETCHING_DATA = 'fetching_data',
	SET_ERROR = 'set_error'
}

export default BoardActions